refactor: 로그인/로그아웃 로직을 AuthContext로 이동하여 구조 정리 #20
